+++ db/migrate/20230619222337_create_comments.rb
@@ -2,9 +2,9 @@ class CreateComments < ActiveRecord::Migration[6.1]
   def change
     create_table :comments do |t|
 
-      t.integer :customer_id, null: false
-      t.integer :comment_genre_id, null: false
-      t.integer :food_id, null: false
+      t.bigint :customer_id, null: false
+      t.bigint :comment_genre_id, null: false
+      t.bigint :food_id, null: false
       t.text :comment, null: false
 
       t.timestamps
