+++ db/migrate/20230619222441_create_recipes.rb
@@ -2,7 +2,7 @@ class CreateRecipes < ActiveRecord::Migration[6.1]
   def change
     create_table :recipes do |t|
 
-      t.integer :customer_id, null: false
+      t.bigint :customer_id, null: false
       t.string :name, null: false
       t.text :process
       t.text :comment
