+++ db/migrate/20230619222453_create_recipe_details.rb
@@ -2,8 +2,8 @@ class CreateRecipeDetails < ActiveRecord::Migration[6.1]
   def change
     create_table :recipe_details do |t|
 
-      t.integer :recipe_id, null: false
-      t.integer :food_id, null: false
+      t.bigint :recipe_id, null: false
+      t.bigint :food_id, null: false
       t.integer :amount, null: false
 
       t.timestamps
