+++ db/migrate/20230627112340_create_cutting_board_foods.rb
@@ -2,8 +2,8 @@ class CreateCuttingBoardFoods < ActiveRecord::Migration[6.1]
   def change
     create_table :cutting_board_foods do |t|
 
-      t.integer :customer_id, null: false
-      t.integer :food_id, null: false
+      t.bigint :customer_id, null: false
+      t.bigint :food_id, null: false
       t.integer :amount, null: false
 
       t.timestamps
